+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3 (unreleased)
 
+- Added `--require-files` option
 - Improved output when multiple versions of the same dependency
 
 ## 0.1.2 (2022-11-30)